Reduce Worker indirection for bundlers
Now that wasm-bindgen has `wasm_bindgen::link_to!`, we can use it to
ensure that Worker is emitted alongside the snippet file instead of
having to use the snippet file itself as a Worker source.

In turn, this allows to use static import of the main module in the
Worker, which avoids unnecessary extra code splitting we used to have.

Aside from slightly better chunking, this also happens to work around
the Parcel issue https://github.com/parcel-bundler/parcel/issues/8727.